Better handle of ebooks with empty title
Due to an error in a SQL script, in Mannheim, are deleted title contains in the csv file. They have no title! These records are now removed with a better algorithm from the list, a special case code removed for this case. These titles are also not listed in the error file.